Added Dependency Injection into the application. Now the Network Portion of the app is all DI into the app and I no longer need to create a Network Service.
Added FurballApp and added the AppComponent to create all of the modules.

The AppComponent holds all of the modules.
RetrofitModule - Builds Retrofit
RestApiModule - links the retrofit/webApi to the WebRestApi.

Now in the ImagesFragment on create we inject the Modules into the fragment, from there we now use the new WebRestApi to call the webservice to get the data from the server.