Added a check to see if the images fragment has already been created (eg on a screen rotation) and loads the previous fragment if it already exists otherwise creates a new fragment.
ImagesFragment Added a check to check if data is already been loaded from the server and if it has use that data instead of the going and getting a new set of data.

Made the WebResponse Object Parcelable so that it can be easily passed in a bundle.

Created a Test case to make sure the WebResponse parcel is being created correctly.

removed android:configChanges="orientation" from the MainActivity tag because it was never supposed to be there.